--- app/routes/register-your-interest/register-your-interest.js
@@ -1,15 +1,0 @@
-const urlPrefix = require('../../config/index').urlPrefix
-const ruralPaymentsAgency = require('../../config/index').ruralPaymentsAgency
-
-module.exports = [
-  {
-    method: 'GET',
-    path: `${urlPrefix}/register-your-interest`,
-    options: {
-      auth: false,
-      handler: async (request, h) => {
-        return h.view('register-your-interest/register-your-interest', { ruralPaymentsAgency })
-      }
-    }
-  }
-]
